Big refactor! Pulled components out!
Tag index and pages for each tag, generated via a script! Tag index
shows how many posts there are for each tag. woo!

We use catch-links node module to stay within the SPA even if showing
markdown-generated HTML.

Discovered that the babel used by gatsby is 5.x and the latest is 6.x
and they have completely incompatible config. So running 
generate_tags.js is a total pain. So much of a pain that I wrote a
shell script to change the babel config whenever we run it. Boo.

404 works as a catch-all (and I'm pretty sure) as a built 404.html file,
so it will give me what I'm looking for. This required a gatsby change
I'll probably PR pretty soon here.