The great upgrade for gatsby 0.8.x
Updated react-router, different method for programmatic navigation.

Things are no longer passed by props:
 - page/pages are passed on the route
 - config is to be required directly

Changed and upgraded dependencies.

Lodash is the big one. No more '/collection/' in specific requires, so
that's nice. Would have waited to upgrade, but a number of the other
dependencies needed it, thought it would be easier to upgrade all at
once.