Move to more reasonable CSS plan
We get hot reload back in dev mode
We use a separate, cacheable, styles.css in production

The difference is that we always generate bundle.js. We need to because
today it's generated only if bundle.js is generated as well. A slight 
slowdown, a bundle.js file in public/ and shipped to my servers, but
no big problem.