refactor: Move Author to Wrapper, check for null title/path
Also, consolidate function to go from post to components in index,
popular and tags components.

The move of Author to wrapper means that it can be deleted at the bottom
of all six page components. As part of is, also removed unneeded plain
wrapper divs.

Feels good to delete code!